♻️ refactor: Create create_methods module to assist App class
